minarc 1.0.30 / orc singing out the graveyard

+++ code/orc/updateOrcProduct.rb
@@ -3,10 +3,16 @@
 # == Synopsis
 #
 # This is an NRTP Orchestrator command line tool used to update the statuses of NRTP jobs.
+#
 # 
 # -f flag:
 #
 # Mandatory flag. This option is used to specify the name of the file to be updated.  
+#
+#
+# -i flag:
+#
+# Alternative to "-f" mandatory flag. This option is used to specify the job-order-id.  
 #
 #
 # -s flag:
@@ -16,19 +22,32 @@
 # FAILURE
 # SUCCESS
 # OBSOLETE
+# DISCARDED
+#
+# -a flag and -b flag:
+#
+# Optional flag. -a and -b must go together.
+# -a is the starting time 
+# -b is the stoping time.
+# Its used to locate and update the same filename but with different sensing times
+# 
 #
 #
 # == Usage
-# updateOrcProduct.rb -f <file-name> -s <new-status>
+# updateOrcProduct.rb -f <file-name> | -i <job-id>  -s <new-status>
 #
 #     --file <file-name>         specifies the name of the file to be updated
+#     --id <job-order-id>        specifies the job-order-id to be used
 #     --set-to                   specifies the new status of the NRTP job
 #     --delete                   deletes all
 #     --help                     shows this help
 #     --usage                    shows the usage
 #     --Debug                    shows Debug info during the execution
 #     --version                  shows version number
+#     --start                    it specifies a starting sensing time to be archived with. Shortcut = -a
+#     --stop                     it specifies a stoping sensing time to be archived with. Shortcut = -b
 # 
+#
 # == Author
 # DEIMOS-Space S.L.
 #
@@ -40,7 +59,7 @@
 #
 # === SMOS NRTP Orchestrator
 #
-# CVS: $Id: updateOrcProduct.rb,v 1.6 2008/07/29 14:52:20 decdev Exp $
+# CVS: $Id: updateOrcProduct.rb,v 1.8 2009/03/18 14:33:41 algs Exp $
 #
 #########################################################################
 
@@ -51,7 +70,7 @@ require "orc/ORC_DataModel"
 require "cuc/EE_ReadFileName"
 
 # Global variables
-@@dateLastModification = "$Date: 2008/07/29 14:52:20 $"   # to keep control of the last modification
+@@dateLastModification = "$Date$"   # to keep control of the last modification
                                        # of this script
                                        # execution showing Debug Info
 
@@ -60,10 +79,11 @@ require "cuc/EE_ReadFileName"
 def main
 
    #Hardcoded values
-   @arrStatus = ["SUC", "FAI", "OBS"]
+   @arrStatus = ["SUC", "FAI", "OBS", "DIS"]
 
    # Data provided by the user
    @filename               = ""
+   @id                     = 0
    @newStatus              = ""
    @isDebugMode            = false
    @bDeleteAll             = false
@@ -71,6 +91,12 @@ def main
    @bDeleteFailed          = false
    @bDeleteSuccess         = false
    @bDeleteQueued          = false
+   @bDeleteDiscarded       = false
+   @bFileOrId              = false
+   @start                  = ""
+   @stop                   = ""
+   @bStart                 = false
+   @bStop                  = false
 
 #    # Data generated or extracted from filename
 #    # (for the generated product)
@@ -87,7 +113,10 @@ def main
    
    opts = GetoptLong.new(
      ["--file",       "-f",       GetoptLong::REQUIRED_ARGUMENT],
+     ["--id",         "-i",       GetoptLong::REQUIRED_ARGUMENT],
      ["--set-to",     "-s",       GetoptLong::REQUIRED_ARGUMENT],
+     ["--start",      "-a",       GetoptLong::REQUIRED_ARGUMENT],
+     ["--stop",       "-b",       GetoptLong::REQUIRED_ARGUMENT],
      ["--Debug",      "-D",       GetoptLong::NO_ARGUMENT],
      ["--delete",     "-d",       GetoptLong::NO_ARGUMENT],
      ["--usage",      "-u",       GetoptLong::NO_ARGUMENT],
@@ -96,6 +125,7 @@ def main
      ["--SUCCESS",    "-S",       GetoptLong::NO_ARGUMENT],
      ["--QUEUED",     "-Q",       GetoptLong::NO_ARGUMENT],
      ["--OBSOLETE",   "-O",       GetoptLong::NO_ARGUMENT],
+     ["--DISCARDED",  "-I",       GetoptLong::NO_ARGUMENT],
      ["--help",       "-h",       GetoptLong::NO_ARGUMENT]
      )
     
@@ -104,17 +134,25 @@ def main
          case opt      
             when "--Debug"     then @isDebugMode = true
             when "--version" then	    
-               print("\nESA - DEIMOS-Space S.L. ", File.basename($0), " $Revision: 1.6 $  [", @@dateLastModification, "]\n\n\n")
+               print("\nESA - DEIMOS-Space S.L. ", File.basename($0), " $Revision: 1.8 $  [", @@dateLastModification, "]\n\n\n")
                exit(0)
-            when "--FAILED"        then @bDeleteFailed       = true
-            when "--SUCCESS"       then @bDeleteSuccess      = true
-            when "--QUEUED"        then @bDeleteQueued       = true
-            when "--OBSOLETE"      then @bDeleteObsolete     = true
-	         when "--delete"        then @bDeleteAll          = true
-            when "--file"          then @filename            =  arg.to_s
-            when "--set-to"        then @newStatus           = (arg.to_s).upcase[0..2]
-			   when "--help"          then RDoc::usage
-	         when "--usage"         then RDoc::usage("usage")
+            when "--start"          then @start                = arg.to_s 
+                                         @bStart               =  true 
+            when "--stop"           then @stop                 = arg.to_s 
+                                         @bStop                =  true     
+            when "--FAILED"         then @bDeleteFailed        = true
+            when "--SUCCESS"        then @bDeleteSuccess       = true
+            when "--QUEUED"         then @bDeleteQueued        = true
+            when "--OBSOLETE"       then @bDeleteObsolete      = true
+            when "--DISCARDED"      then @bDeleteDiscarded     = true
+	         when "--delete"         then @bDeleteAll           = true
+            when "--file"           then @filename             = arg.to_s
+                                         @bFileOrId            = true
+            when "--id"             then @id                   = arg
+                                         @bFileOrId            = false
+            when "--set-to"         then @newStatus            = (arg.to_s).upcase[0..2]
+			   when "--help"           then RDoc::usage
+	         when "--usage"          then RDoc::usage("usage")
          end
       end
    rescue Exception
@@ -125,24 +163,48 @@ def main
    if @bDeleteAll == true then
       deleteAll
       exit
-   end   
-
-
-   if @filename == "" or @filename.length > 100 then
-      puts
-      puts
-      RDoc::usage("usage")
-      puts
+   end 
+    
+   if (@bStart == true and @bStop == false) or (@bStart == false and @bStop == true) then
+      puts "-start and -stop options must go together"
       exit(99)
-   end
+   end  
 
-   #check that the file is registered as a trigger product
-   @triggerProd = TriggerProduct.find_by_filename(@filename)
+   if @bFileOrId == true then
+      if @filename == "" or @filename.length > 100 then
+         puts
+         puts
+         RDoc::usage("usage")
+         puts
+         exit(99)
+      end
+      #check that the file is registered as a trigger product (by filename)
+      if @bStart == true then
+         @triggerProd = TriggerProduct.find_by_filename_and_sensing_start_and_sensing_stop(@filename, @start, @stop)
+      else
+         @triggerProd = TriggerProduct.find_by_filename(@filename)    
+      end  
+   else
+      if @id == 0 then
+         puts
+         RDoc::usage("usage")
+         puts
+         exit(99)
+      end
+      #check that the file is registered as a trigger product (by Id)
+      if @bStart == true then
+         @triggerProd = TriggerProduct.find_by_id_and_sensing_start_and_sensing_stop(@id, @start, @stop)
+      else
+         @triggerProd = TriggerProduct.find_by_id(@id)
+      end
+   end
+   
+
    if @triggerProd == nil then
       puts
-      puts "The filename provided is not registered as a trigger product :-("
+      puts "#{@filename} is not registered as a trigger product :-("
       puts
-      RDoc::usage("usage")
+      # RDoc::usage("usage")
       puts
       exit(99)
    else
@@ -264,6 +326,11 @@ def main
       exit(0)
    end
 
+   if @newStatus == "DIS" then
+      set_product_to_discarded
+      exit(0)
+   end
+
 end
 
 #-------------------------------------------------------------------------------
@@ -324,7 +391,7 @@ end
 #-------------------------------------------------------------------------------
 def set_product_to_failure
 
-   #check that the file is not allready present in the failing products list
+   # Check that the file is not already present in the failing products list
    tmpProd = FailingTriggerProduct.find_by_trigger_product_id(@trigger_prod_id)
    if tmpProd != nil then
       puts
@@ -338,9 +405,9 @@ def set_product_to_failure
 
    OrchestratorQueue.transaction do
 
-      tmpProd = FailingTriggerProduct.new
-      tmpProd.trigger_products = @triggerProd
-      tmpProd.failure_date = Time.now
+      tmpProd                    = FailingTriggerProduct.new
+      tmpProd.trigger_products   = @triggerProd
+      tmpProd.failure_date       = Time.now
       
       begin
         tmpProd.save!
@@ -365,6 +432,39 @@ def set_product_to_failure
       puts "update completed :-)"
       puts
    end
+
+   # Other pending jobs with same trigger-file (different time window)
+   # are "moved" to DISCARDED
+   
+   arrStillQueued = OrchestratorQueue.getAllQueuedByName(@triggerProd.filename)
+
+   if @isDebugMode == true then
+      puts "Discarding : "
+      arrStillQueued.each{|queuedFile|
+         puts queuedFile.id
+      }
+   end
+
+   arrStillQueued.each{|queuedFile|
+      
+      tmpProd = DiscardedTriggerProduct.new
+      tmpProd.trigger_products   = queuedFile
+      tmpProd.discarded_date     = Time.now
+
+      begin
+        tmpProd.save!
+      rescue Exception => e
+         puts e.to_s
+         puts
+         puts "Unable to register #{queuedFile.filename} as a DISCARDED trigger product :-("
+         puts
+         puts
+         exit(99)
+      end
+
+      OrchestratorQueue.delete_all(:trigger_product_id => queuedFile.id)
+            
+   }
  
 end
 
@@ -419,6 +519,58 @@ end
 
 #-------------------------------------------------------------------------------
 
+
+def set_product_to_discarded
+
+   #check that the file is not allready present in the discarded products list
+   tmpProd = DiscardedTriggerProduct.find_by_trigger_product_id(@trigger_prod_id)
+
+   if tmpProd != nil then
+      puts
+      puts "The trigger product is registered in both 'orchestrator queue' and 'discarded trigger products' :-("
+      puts "Incoherent orchestrator status !"
+      puts
+      RDoc::usage("usage")
+      puts
+      exit(99)
+   end
+
+   OrchestratorQueue.transaction do
+
+      disTime = Time.now
+      tmpProd = DiscardedTriggerProduct.new
+      tmpProd.trigger_products = @triggerProd
+      tmpProd.discarded_date      = disTime
+
+      begin
+         tmpProd.save!
+      rescue
+         puts
+         puts "Unable to register #{@filename} as a discarded trigger product :-("
+         puts "Unknown error !"
+         puts
+         exit(99)
+      end
+
+      OrchestratorQueue.delete_all(:trigger_product_id => @trigger_prod_id)
+
+   end
+
+   if @isDebugMode then
+      puts
+      puts "updateOrcProduct => update status to 'discarded trigger product'..."
+      puts "Product name     : #{@filename}"
+      puts "Product id       : #{@trigger_prod_id}"
+      puts "Discarded since  : #{obsTime}"
+      puts "update completed :-)"
+      puts
+   end
+ 
+end
+
+#-------------------------------------------------------------------------------
+
+
 def deleteAll
    if @bDeleteQueued == true then
       OrchestratorQueue.delete_all
@@ -439,8 +591,13 @@ def deleteAll
       FailingTriggerProduct.delete_all
       puts "orchestrator failed trigger files deleted"
    end
+
+   if @bDeleteDiscarded == true then
+      DiscardedTriggerProduct.delete_all
+      puts "orchestrator discarded trigger files deleted"
+   end
    
-   if @bDeleteQueued == true and @bDeleteFailed == true and @bDeleteSuccess == true and @bDeleteObsolete == true then
+   if @bDeleteQueued == true and @bDeleteFailed == true and @bDeleteSuccess == true and @bDeleteObsolete == true and @bDeleteDiscarded then
       TriggerProduct.delete_all
    end
 end
